R2_Sim_Cont_Cnt_Bin_Simple_24MAY2016
This code does not reproduce the results in the paper 'An R Squared for Fixed effects in the GLMM,' but it is much easier to run because it doesn't require SAS. The results shown here use the residual degrees of freedom to compute R squared beta. To use the Kenward Roger approach for the GLMM R squared beta, the SAS macro must be used.